requirements.txt
MetaTrader5==5.0.4993
fastapi==0.115.2
uvicorn==0.32.0
pydantic==2.8.2
python-dotenv==1.0.0
schedule==1.2.0
pytest==7.4.0